// todo-angular/src/openapi-services/model/detailedTodoDTO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DetailedTodoDTO { 
    description?: string;
    done?: boolean;
    id?: number;
    title?: string;
}